requirements.txt
flask
python-dotenv
Pillow
reportlab
requests
